@@ -1,12 +1,8 @@
-// backend/routes/auth.js
 const express = require("express");
+const { registerUser, loginUser } = require("../controllers/authController");
 const router = express.Router();
-const { register, login } = require("../authController");  // ✅ correct path
 
-// Register route
-router.post("/register", register);
-
-// Login route
-router.post("/login", login);
+router.post("/register", registerUser);
+router.post("/login", loginUser);
 
 module.exports = router;
