@@ -1,78 +1,97 @@
 // backend/server.js
-require("dotenv").config();
 const express = require("express");
-const cors = require("cors");
+const dotenv = require("dotenv");
 const http = require("http");
-const { Server } = require("socket.io");
-const connectDB = require("./config/db");
+const cors = require("cors");
+const path = require("path");
+const connectDB = require("./config/db.js");
+
+dotenv.config();
 
 const app = express();
 
-// ---------- Middlewares ----------
+// ------------------- Middleware -------------------
 app.use(express.json());
-app.use(
-  cors({
-    origin: "*",
-    credentials: true,
-    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
-    allowedHeaders: ["Content-Type", "Authorization"],
-  })
-);
-app.options("*", cors());
+app.use(cors());
 
-// ---------- Routes ----------
-app.use("/api/auth", require("./routes/auth"));
-app.use("/api/users", require("./routes/users.routes")); // already in your project
-app.use("/api/messages", require("./routes/messages.routes")); // ⬅️ NEW
+// ------------------- Database -------------------
+connectDB();
 
-// Health check
-app.get("/api/health", (_req, res) => res.json({ status: "ok" }));
-
-// Global Error Handler (keep last)
-app.use((err, _req, res, _next) => {
-  console.error("🔥 Server Error:", err);
-  res.status(500).json({ message: "Internal server error" });
-});
-
-// ---------- HTTP + Socket.IO ----------
-const server = http.createServer(app);
-const io = new Server(server, {
-  cors: { origin: "*", methods: ["GET", "POST"] },
-});
-
-io.on("connection", (socket) => {
-  console.log("🔌 User connected:", socket.id);
-
-  // Optional: private rooms now or later
-  socket.on("join_room", (room) => {
-    socket.join(room);
-    console.log(`📌 ${socket.id} joined room: ${room}`);
-  });
-
-  socket.on("send_message", ({ room, message, sender }) => {
-    console.log(`💬 [${room}] ${sender}: ${message}`);
-    io.to(room).emit("receive_message", { room, message, sender });
-  });
-
-  socket.on("chat:message", (msg) => {
-    io.emit("chat:message", msg);
-  });
-
-  socket.on("disconnect", () => {
-    console.log("❌ User disconnected:", socket.id);
-  });
-});
-
-// ---------- Start ----------
-const PORT = process.env.PORT || 5000;
-(async () => {
+// ------------------- Route Loader -------------------
+/**
+ * Mounts Express router with robust checks.
+ * Logs success or failure for each route.
+ */
+const mountRoute = (mountPath, modulePath) => {
   try {
-    await connectDB();
-    server.listen(PORT, "0.0.0.0", () =>
-      console.log(`🚀 Server running at http://0.0.0.0:${PORT}`)
-    );
+    const resolved = path.resolve(__dirname, modulePath);
+    const mod = require(resolved);
+
+    let candidate = mod;
+    if (mod && typeof mod === "object" && mod.default) candidate = mod.default;
+    if (candidate && typeof candidate === "object" && candidate.router)
+      candidate = candidate.router;
+
+    const isRouter =
+      typeof candidate === "function" ||
+      (candidate &&
+        (typeof candidate.use === "function" ||
+          typeof candidate.handle === "function" ||
+          Array.isArray(candidate.stack)));
+
+    if (isRouter) {
+      app.use(mountPath, candidate);
+      console.log(`✅ Mounted route ${mountPath} -> ${modulePath}`);
+    } else {
+      console.warn(`⚠️ Route module ${modulePath} did not export an Express router.`);
+    }
   } catch (err) {
-    console.error("❌ Failed to start server:", err);
-    process.exit(1);
+    console.error(`❌ Failed to mount route ${modulePath}:`, err.message);
   }
-})();
+};
+
+// ------------------- API Routes -------------------
+mountRoute("/api/auth", "./routes/auth");
+mountRoute("/api/users", "./routes/users.routes");
+mountRoute("/api/messages", "./routes/messages.routes");
+
+// Health check routes
+app.get("/", (req, res) => res.send("✅ API is running..."));
+app.get("/health", (req, res) => res.json({ ok: true }));
+
+// ------------------- Create HTTP Server -------------------
+const server = http.createServer(app);
+
+// ------------------- Socket.IO Integration -------------------
+try {
+  require("./socket")(server, app);
+  console.log("✅ Socket.IO attached successfully");
+} catch (err) {
+  console.error("❌ Failed to attach Socket.IO:", err.message);
+}
+
+// ------------------- Start Server -------------------
+const PORT = process.env.PORT || 5000;
+server.listen(PORT, () => console.log(`🚀 Server running on port ${PORT}`));
+
+// ------------------- Global Error Handlers -------------------
+process.on("unhandledRejection", (reason, promise) => {
+  console.error("💥 Unhandled Rejection at:", promise, "reason:", reason);
+});
+
+process.on("uncaughtException", (err) => {
+  console.error("💥 Uncaught Exception:", err);
+  // optional: process.exit(1);
+});
+
+// Optional: graceful shutdown on SIGTERM/SIGINT
+process.on("SIGTERM", () => {
+  console.log("🔹 SIGTERM received. Closing server...");
+  server.close(() => process.exit(0));
+});
+
+process.on("SIGINT", () => {
+  console.log("🔹 SIGINT received. Closing server...");
+  server.close(() => process.exit(0));
+});
+
