@@ -1,2 +1,2 @@
 // Use Android emulator loopback for backend
-export const API_URL = 'http://10.0.2.2:5000';
+export const API_URL = 'http://192.168.43.176:5000';
